Refactor Resource to load retrieval strategy as class loads
