+++ lib/jsonapi-resources.rb
@@ -4,6 +4,7 @@ require 'jsonapi/compiled_json'
 require 'jsonapi/active_relation_retrieval'
 require 'jsonapi/active_relation_retrieval_v09'
 require 'jsonapi/active_relation_retrieval_v10'
+require 'jsonapi/resource_common'
 require 'jsonapi/resource'
 require 'jsonapi/cached_response_fragment'
 require 'jsonapi/response_document'
